# backend/src/Cv.Application/Courses/CourseDto.cs
using Cv.Domain.Courses;

namespace Cv.Application.Courses;

public sealed record CourseOverviewDto(decimal TotalCredits, IReadOnlyList<CourseDto> Courses);

public sealed record CourseDto(
    string Name,
    decimal Credits,
    CourseLevel Level,
    DateOnly CompletedOn,
    bool Highlighted);